requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
anthropic>=0.40.0
openai>=1.0.0
pandas>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
